Seems can run?

+++ model/diffusion/modules.py
@@ -10,7 +10,7 @@ class SinusoidalPosEmb(tf.keras.layers.Layer):
         half_dim = self.dim // 2
         emb = math.log(10000) / (half_dim - 1)
         emb = tf.exp(tf.range(half_dim, dtype=tf.float32) * -emb)
-        emb = x[:, None] * emb[None, :]
+        emb = tf.cast(x[:, None], tf.float32) * emb[None, :]
         emb = tf.concat([tf.sin(emb), tf.cos(emb)], axis=-1)
         return emb
